merge feature-recommendation into main

+++ backend/src/main/java/com/example/video/controller/FeedController.java
@@ -19,7 +19,7 @@ import java.util.UUID;
 @RequestMapping("/api/feed")
 @CrossOrigin(origins = "*")
 public class FeedController {
-
+ 
     @Autowired
     private RecommendationService recommendationService;
 
@@ -50,13 +50,12 @@ public class FeedController {
     @PostMapping("/view/{videoId}")
     public ResponseEntity<Map<String, Object>> recordView(
             @PathVariable UUID videoId,
-            @RequestParam(defaultValue = "0") int watchDuration,
+            @RequestParam(defaultValue = "0") float watchDuration,
             @RequestParam(defaultValue = "false") boolean completed,
             Authentication authentication) {
         
         UUID currentUserId = getCurrentUserId(authentication);
         socialService.recordView(videoId, currentUserId, watchDuration, completed);
-        
         Map<String, Object> response = new HashMap<>();
         response.put("success", true);
         response.put("message", "View recorded");
